requirements.txt
reportlab==4.2.2
svglib==1.5.1
Pillow==10.4.0
qrcode==7.4.2
pdf2image>=1.16.0
pypdf>=4.0.0
requests==2.32.3
pydantic>=2.7,<3
PyYAML>=6.0
python-dotenv>=1.0.0
watchdog>=6.0.0
numpy>=2.0.0